Return errors as per JSON API suggestion
See http://jsonapi.org/format/#errors
